Revert "Start session in background so method can return"
This reverts commit 8ccbadba65e9aa7604052f00321c2e841bdc1da7.

Turns out we don't actually need to start the session in the background,
and since on the AAM client we want to know when the session has
launched/failed to launch (to provide feedback) it is actually better
not to. There would be alternative ways to track this but just upping
the daemon request timeout for this method call and waiting for it to
complete seems the simplest and least likely to run into problems.

Just reverting this commit rather than deleting it entirely in case we
need to change this again later.