Fix detection and running of qdesktop
Rather than detecting if qdesktop is available by testing the command
itself, which now doesn't seem to work in a real environment on AWS (if
it ever did), we now test that the binary is present at the usual
location and then run this directly.

This is likely to break in different situations to the previous method,
such as if the binary moves, but since the existing method was broken
already and I can't see a robust way to fix it right now then this is
better for now. This code is likely to change again sooner or later
anyway since this will only work at the moment for SGE clusters, but we
now support Slurm, Torque etc.