Remove reference config files
These are not actually used, they are just here for reference. I find
this a bit confusing, as without checking the deployment script I
assumed they were used and modified them. They also have the potential
to get out of sync with the actual config files. For these reasons I'm
going to remove them from this repo.

Dear people/me of the future, if you're looking for the actual config
files you should look in
https://github.com/alces-software/clusterware-services/tree/master/alces-access-manager-daemon.